# static/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Annoto API",
        "version": "v1"
    },
    "paths": {
        "/tasks/": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get Tasks",
                "description": "Get a list of the IDs of all available labelling tasks",
                "operationId": "get_tasks",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    "ecg-qrs-classification-physiodb"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/tasks/{task_id}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get Task",
                "description": "Get all information about a labelling task",
                "operationId": "get_task",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Task Id",
                            "type": "string"
                        },
                        "example": "ecg-qrs-classification-physiodb",
                        "name": "task_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found!"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/tasks/{task_id}/next": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get Next Image",
                "description": "Get the image that should be annotated",
                "operationId": "get_next_image",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Task Id",
                            "type": "string"
                        },
                        "example": "ecg-qrs-classification-physiodb",
                        "name": "task_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "ecg_1.png"
                            }
                        }
                    },
                    "404": {
                        "description": "No more images to annotate"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/tasks/{task_id}/{src}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get Image",
                "description": "Get the image that should be annotated",
                "operationId": "get_image",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Task Id",
                            "type": "string"
                        },
                        "example": "ecg-qrs-classification-physiodb",
                        "name": "task_id",
                        "in": "path"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "Src",
                            "type": "string"
                        },
                        "example": "sloth.jpg",
                        "name": "src",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "image/*": {
                                "schema": {
                                    "type": "file",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "File not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Save Annotation",
                "description": "Saves the annotation for the specified image",
                "operationId": "save_annotation",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Task Id",
                            "type": "string"
                        },
                        "example": "ecg-qrs-classification-physiodb",
                        "name": "task_id",
                        "in": "path"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "Src",
                            "type": "string"
                        },
                        "example": "sloth.jpg",
                        "name": "src",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnnotationData"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "404": {
                        "description": "File not found!"
                    },
                    "400": {
                        "description": "Hash values of the annotation and the local source do not match!"
                    },
                    "428": {
                        "description": "Provided proofs are not valid!"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    },
                    {
                        "JWTBearer": []
                    }
                ]
            }
        },
        "/users/{user_id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Read User",
                "description": "Read a user by their ID",
                "operationId": "read_user",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "User Id",
                            "type": "integer"
                        },
                        "example": 1,
                        "name": "user_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create User",
                "description": "Create a user",
                "operationId": "create_user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email already registered"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login/": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Login",
                "description": "Validate login via email and password",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Failed to validate login"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ping/": {
            "get": {
                "tags": [
                    "ping"
                ],
                "summary": "Ping",
                "description": "Validate API availability after authorization",
                "operationId": "ping",
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    }
                },
                "security": [
                    {
                        "JWTBearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "AnnotationData": {
                "title": "AnnotationData",
                "required": [
                    "label",
                    "hash",
                    "competency",
                    "is_trained",
                    "is_attentive"
                ],
                "type": "object",
                "properties": {
                    "label": {
                        "title": "Label",
                        "type": "string",
                        "description": "The label the file should be annotated with",
                        "example": "Sloth"
                    },
                    "hash": {
                        "title": "Hash",
                        "type": "string",
                        "description": "The hash of the file",
                        "example": "e922903b4d5431a8f9def3c89ffcb0b18472f3da304f28a2dbef9028b6cd205d"
                    },
                    "competency": {
                        "title": "Competency",
                        "type": "string",
                        "description": "The competencies the annotator has",
                        "example": "Prof. Dr. Med"
                    },
                    "is_trained": {
                        "title": "Is Trained",
                        "type": "boolean",
                        "description": "Whether the annotator said he finished the training",
                        "example": true
                    },
                    "is_attentive": {
                        "title": "Is Attentive",
                        "type": "boolean",
                        "description": "Whether the annotator said that he is attentive",
                        "example": true
                    }
                },
                "description": "The basic Annotation Data for a data file"
            },
            "CreateUserRequest": {
                "title": "CreateUserRequest",
                "required": [
                    "fullname",
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "fullname": {
                        "title": "Fullname",
                        "type": "string",
                        "description": "The full name of the user",
                        "example": "Prof. Dr. Folivora"
                    },
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "description": "The email of the user",
                        "example": "team@folivora.online"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string",
                        "description": "The password of the user",
                        "example": "password"
                    }
                },
                "description": "The fields required to create a new user"
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "LoginRequest": {
                "title": "LoginRequest",
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "description": "The email of the user to be logged in",
                        "example": "team@folivora.online"
                    },
                    "password": {
                        "title": "Password",
                        "type": "string",
                        "description": "The password of the user to be logged in",
                        "example": "password"
                    }
                },
                "description": "The basic Login Data"
            },
            "LoginResponse": {
                "title": "LoginResponse",
                "required": [
                    "access_token"
                ],
                "type": "object",
                "properties": {
                    "access_token": {
                        "title": "Access Token",
                        "type": "string",
                        "description": "The generated JWT access token",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxLCJleHBpcmVzIjoxNjUxNjcyMTU5LjMyMjY3ODZ9.ONVOoCKQ3ypaPgPAYCFnjlECzRdW9QgZsrkilnclnxg"
                    }
                },
                "description": "Login token (JWT)"
            },
            "Task": {
                "title": "Task",
                "required": [
                    "id",
                    "description",
                    "labels"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string",
                        "description": "The identifier of this labelling task",
                        "example": "ecg-qrs-classification-physiodb"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string",
                        "description": "A brief description of this labelling task",
                        "example": "Classifying the QRS complex of some ECGs"
                    },
                    "labels": {
                        "title": "Labels",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "List of labels",
                        "example": [
                            "Atrial fibrillation",
                            "Normal sinus rhythm",
                            "Ventricular tachycardia",
                            "Noise",
                            "Other"
                        ]
                    }
                },
                "description": "A labelling task"
            },
            "UserResponse": {
                "title": "UserResponse",
                "required": [
                    "fullname",
                    "email",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "fullname": {
                        "title": "Fullname",
                        "type": "string",
                        "description": "The full name of the user",
                        "example": "Prof. Dr. Folivora"
                    },
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "description": "The email of the user",
                        "example": "team@folivora.online"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer",
                        "description": "The id of the user",
                        "example": 1
                    }
                },
                "description": "An already existing user"
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "JWTBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}